Revert "Restructure by folders"
This reverts commit 5daa29bfd1a7b73d082f8e014d5bde64280c59bf.
