--- canteen-management-system-client/src/components/LandingPage/LandingPageRight.css
@@ -1,18 +1,0 @@
-.right-container {
-  min-height: 100vh;
-  background-color: #fff;
-}
-
-.right-col {
-  max-width: 400px;
-  width: 100%;
-}
-
-.signin-paragraph.ant-typography {
-  text-align: center !important;
-  margin-top: 16px !important;
-}
-
-.signin-link.ant-typography {
-  color: #3678eb !important;
-}
