--- canteen-management-system-client/src/components/LandingPage/WelcomeBack.css
@@ -1,24 +1,0 @@
-.icon-container {
-  width: 64px;
-  height: 64px;
-  background-color: #1890ff;
-  border-radius: 50%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.icon {
-  font-size: 32px;
-  color: white;
-  margin-bottom: 0px;
-}
-
-.welcome-col {
-  text-align: center;
-  padding: 60px;
-}
-
-.welcome-row {
-  margin-bottom: 20px;
-}
