+++ canteen-management-system-client/src/pages/LandingPage.jsx
@@ -1,6 +1,6 @@
 import LoginLayout from "../layouts/LandingPage";
-import LandingPageLeft from "../components/LandingPage/LandingPageLeft";
-import LandingPageRight from "../components/LandingPage/LandingPageRight";
+import LandingPageLeft from "../components/LandingPageLeft";
+import LandingPageRight from "../components/LandingPageRight";
 
 const LandingPage = () => {
   return (
